pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools>=80.0", "setuptools-scm>=8"]

[project]
name = "python-nsjail"
dynamic = ["version"]
description = "Prebuilt nsjail executables for Python developers"
readme = "README.md"
authors = [{ name = "tanbro", email = "tanbro@163.com" }]
license = "Apache-2.0"
requires-python = ">=3.9"

dependencies = [
  "typing-extension>=3.9.0; python_version < '3.11'",
]

keywords = [
  "sandbox",
  "nsjail",
  "linux",
  "security",
  "process-isolation",
  "chroot",
  "linux-namespaces",
  "seccomp-bpf-policies",
]

classifiers = [
  "Operating System :: POSIX :: Linux",
]

[project.urls]
Homepage = "https://github.com/tanbro/python-nsjail"
Repository = "https://github.com/tanbro/python-nsjail"

[project.scripts]
nsjail = "nsjail.__main__:main"
nsjail-find = "nsjail.find:main"

[tool.setuptools]
include-package-data = true
zip-safe = false

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
nsjail = ["nsjail"]

[tool.setuptools_scm]
version_file = "src/nsjail/version.py"

[dependency-groups]
dev = [
  "setuptools>=80.0",
  "pytest>=8.0",
  "pytest-asyncio",
  "mypy",
]
